Provide CLI to print and delete stale branches

+++ src/index.ts
@@ -1,6 +1,21 @@
 import { deleteLocalBranches, getLocalBranches, getRemoteBranches } from "./utils";
-
+import yargs from 'yargs';
 async function main() {
+    const cli = yargs.options('print-stale-branches', {
+        type: 'boolean',
+        alias: 'p',
+        description: 'Option to print stale branches',
+    }).option('delete-stale-branch', {
+        type: 'boolean',
+        alias: 'd',
+        description: 'Delete stale branches'
+    }).example('git stale-branches -p -d', 'Print and delete stale branches');
+
+    // TODO : check why we are getting type error here
+    const { printStaleBranches, deleteStaleBranch } = cli.argv as { printStaleBranches?: boolean, deleteStaleBranch?: boolean };
+    console.log('print', printStaleBranches, 'delete : ', deleteStaleBranch);
+
+
     const localBranches = getLocalBranches();
     console.log('Local branches : ', localBranches);
 
@@ -12,7 +27,12 @@ async function main() {
         console.log('There is no stale branches found');
         return;
     }
-    deleteLocalBranches(staleBranches);
+    if (printStaleBranches) {
+        console.log(...staleBranches);
+    }
+    if (deleteStaleBranch) {
+        deleteLocalBranches(staleBranches);
+    }
 }
 
 main().catch(e => {
